+++ tests/python/test_continous_dynamics.py
@@ -10,7 +10,7 @@ from utils import finite_diff
 
 space = manifolds.R3() * manifolds.SO3()
 nu = 0
-epsilon = 1e-4
+epsilon = 1e-3
 
 
 class MyODE(dynamics.ODEAbstract):
@@ -88,12 +88,10 @@ def test_multibody_free():
         ode.forward(x0, u0, data)
         ode.dForward(x0, u0, data)
 
-        Jx0 = data.Jx.copy()
-        Ju0 = data.Ju.copy()
         Jxdiff, Judiff = finite_diff(ode, space, x0, u0)
 
-        err_Jx = np.max(Jxdiff - Jx0)
-        err_Ju = np.max(Judiff - Ju0)
+        err_Jx = np.max(Jxdiff - data.Jx)
+        err_Ju = np.max(Judiff - data.Ju)
 
         assert err_Jx < epsilon and err_Ju < epsilon
     except ImportError:
@@ -166,12 +164,10 @@ def test_centroidal_diff():
     ode.forward(x0, u0, data)
     ode.dForward(x0, u0, data)
 
-    Jx0 = data.Jx.copy()
-    Ju0 = data.Ju.copy()
     Jxdiff, Judiff = finite_diff(ode, space, x0, u0)
 
-    assert np.linalg.norm(Jxdiff - Jx0) <= epsilon
-    assert np.linalg.norm(Judiff - Ju0) <= epsilon
+    assert np.linalg.norm(Jxdiff - data.Jx) <= epsilon
+    assert np.linalg.norm(Judiff - data.Ju) <= epsilon
 
 
 def test_continuous_centroidal():
@@ -243,12 +239,10 @@ def test_continuous_centroidal_diff():
     ode.forward(x0, u0, data)
     ode.dForward(x0, u0, data)
 
-    Jx0 = data.Jx.copy()
-    Ju0 = data.Ju.copy()
     Jxdiff, Judiff = finite_diff(ode, space, x0, u0)
 
-    assert np.linalg.norm(Jxdiff - Jx0) <= epsilon
-    assert np.linalg.norm(Judiff - Ju0) <= epsilon
+    assert np.linalg.norm(Jxdiff - data.Jx) <= epsilon
+    assert np.linalg.norm(Judiff - data.Ju) <= epsilon
 
 
 if __name__ == "__main__":